+++ protocols/request-response/CHANGELOG.md
@@ -1,4 +1,8 @@
-# 0.20.0 [unreleased]
+# 0.21.0 [unreleased]
+
+- Update to `libp2p-swarm` `v0.39.0`.
+
+# 0.20.0
 
 - Update to `libp2p-swarm` `v0.38.0`.